refactor(models): move the OA curation shapes out of the asset package (#161)
HostInsight / HostScore / HostRole / HostPriority are the OSINT Analyst's
*curation* of the surface - its judgements about which hosts matter and why -
not OAM asset types. OpenPortsMap is a host-keyed recon-output map, also not
an OAM asset. None belong in models/asset/ (the OAM package).

- Move all five from models/asset/host.py into models/insight.py, which is
  already the OA curation / finalisation home (HostAnnotation, the insight
  validation report, ReconFinalisationError). insight.py imports VulnProperty
  from models.asset for HostInsight.vulns; cycle-safe (nothing in asset /
  finding / h1 imports insight).
- Drop models/asset/host.py; asset/__init__ and its docstring no longer
  claim the curation shapes. models/__init__ re-exports them from
  models.insight; the asset and insight table rows updated.
- models/attack/graph.py imports HostInsight from models.insight now.
- tests/models/asset/test_host.py -> tests/models/test_insight.py.

Every consumer uses 'from models import X', so no import sites changed
outside the models package. Pure relocation, no shape or behaviour change.
