+++ models/asset/__init__.py
@@ -1,15 +1,17 @@
 """
 models.asset - the recon-output / OAM inventory shapes (OSINT Analyst -> PT).
 
-What the OSINT Analyst's sweep / annotation / finalisation produces:
-endpoints discovered, hostnames classified by role and priority, open
-ports per host, LLM-backed endpoint flags, the OAM asset shapes
-(``Service`` / ``Product`` / ``ProductRelease`` / ``TLSCertificate`` /
-``IpAsset`` / the registrant shapes in ``network``), and the
-``VulnProperty`` annotations hung off them. The bundle that wraps the lot
-for downstream agents - ``AttackGraph`` - lives in ``models.attack`` with
-its ``AttackTree`` / ``AttackForest`` siblings: it composes these shapes but
-is not itself an OAM asset.
+The OAM asset shapes the OSINT Analyst's sweep emits and the PT consumes:
+endpoints discovered, the ``Service`` / ``Product`` / ``ProductRelease`` /
+``TLSCertificate`` / ``IpAsset`` assets and the registrant shapes in
+``network``, plus the ``VulnProperty`` annotations hung off them and the
+LLM-backed endpoint marker.
+
+Two things deliberately live *outside* this package because they are not OAM
+asset types: the OA's *curation* shapes (``HostInsight`` / ``HostScore`` /
+``HostRole`` / ``HostPriority`` / ``OpenPortsMap``) in ``models.insight``,
+and the bundle that wraps the lot - ``AttackGraph`` - in ``models.attack``
+with its ``AttackTree`` / ``AttackForest`` siblings.
 
 Promoted from a single ``models/asset.py`` to a package as the OAM asset
 layer grew. One module per cohesive OAM-asset concern; this ``__init__`` is
@@ -25,8 +27,6 @@ through the scope filter.
 |---|---|
 | ``models.asset.vuln`` | ``VulnProperty`` |
 | ``models.asset.endpoint`` | ``Endpoint``, ``EndpointPage``, ``LlmEndpoint`` |
-| ``models.asset.host`` | ``HostRole``, ``HostPriority``, ``HostInsight``, |
-|                       | ``HostScore``, ``OpenPortsMap`` |
 | ``models.asset.service`` | ``Service``, ``Product``, ``ProductRelease`` |
 | ``models.asset.certificate`` | ``TLSCertificate`` |
 | ``models.asset.ip`` | ``IpAsset`` |
@@ -34,21 +34,14 @@ through the scope filter.
 |                          | ``RdapRecord``, ``DomainRecord`` |
 
 The intra-package import order is a DAG: ``vuln`` / ``certificate`` /
-``network`` are leaves; ``endpoint`` / ``host`` / ``service`` / ``ip`` build
-on them. No cycles, so no ``model_rebuild`` is needed.
+``network`` are leaves; ``endpoint`` / ``service`` / ``ip`` build on them.
+No cycles, so no ``model_rebuild`` is needed.
 """
 
 from __future__ import annotations
 
 from models.asset.certificate import TLSCertificate
 from models.asset.endpoint import Endpoint, EndpointPage, LlmEndpoint
-from models.asset.host import (
-    HostInsight,
-    HostPriority,
-    HostRole,
-    HostScore,
-    OpenPortsMap,
-)
 from models.asset.ip import IpAsset
 from models.asset.network import (
     AsnRecord,
@@ -67,13 +60,8 @@ __all__ = [
     "DomainRecord",
     "Endpoint",
     "EndpointPage",
-    "HostInsight",
-    "HostPriority",
-    "HostRole",
-    "HostScore",
     "IpAsset",
     "LlmEndpoint",
-    "OpenPortsMap",
     "Product",
     "ProductRelease",
     "RdapRecord",
